requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
pandas==2.2.3
boto3==1.36.16
requests==2.32.3
cryptography==44.0.0
PyJWT==2.10.1
gunicorn==23.0.0
transformers==4.48.3
torch
openpyxl
pytz
